Physically move integration-tier tests into test/integration/, not just config-excluded
