+++ packages/claude-sdk-tools/test/integration/TsSpawnCwd.spec.ts
@@ -2,7 +2,7 @@ import { mkdtempSync, rmSync, writeFileSync } from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
-import type { TsServerBridge } from '../src/typescript/TsServerBridge';
+import type { TsServerBridge } from '../../src/typescript/TsServerBridge';
 import { buildTsBridge } from './buildTsBridge';
 
 describe('tsserver spawn cwd is inert', () => {
