+++ packages/claude-sdk-tools/test/integration/fixtures/ts-project/main.ts
