+++ packages/claude-sdk-tools/test/integration/fixtures/ts-project/tsconfig.json
