+++ packages/claude-sdk-tools/test/integration/pipeline-teardown.spec.ts
