+++ packages/claude-sdk-tools/vitest.config.ts
@@ -1,16 +1,14 @@
 import { defineConfig } from 'vitest/config';
 
-// TsServerBridge specs spawn a real tsserver child process; ExecV3/pipeline-teardown.spec.ts
-// spawns real pipe/SIGPIPE behaviour no fake can produce. Both are real OS resources with no
-// substitute for the thing under test, kept out of the default run entirely (see
-// vitest.integration.config.ts) so `vitest`/`pnpm test` never touches them.
-const INTEGRATION_SPECS = ['test/Ts*.spec.ts', 'test/ExecV3/pipeline-teardown.spec.ts'];
-
+// test/integration/ holds specs that spawn a real tsserver process or real pipe/SIGPIPE
+// behaviour no fake can produce — real OS resources with no substitute for the thing under
+// test. Kept out of the default run entirely, physically and logically (see
+// vitest.integration.config.ts), so `vitest`/`pnpm test` never touches them.
 export default defineConfig({
   test: {
     testTimeout: 10_000,
     maxWorkers: '50%',
     include: ['test/**/*.spec.ts'],
-    exclude: INTEGRATION_SPECS,
+    exclude: ['test/integration/**'],
   },
 });
